we will use jsonnet instead from this commit on
the old "flux/" directory is replaced by "manifests", and the manifest
generation workflow takes place on any commit we tagged in "release" as
a release tag, or as a commit that is merged on the release branch.

(Protect your release branch with eg. branch policies as it represents
the production deployment here.)

The CI user should be allowed to push directly to the deploy branch; it
also represents a production environment here and must be protected too.