Express a duplicated structure better with jsonnet
With this new structure, it is made clear the role of "test" and "prod"
strings, and how additional environments could be added eg. "stg", "dev"
without necessarily creating more such structural duplication

We use an 'if' statement to preserve the property that test and prod
have subtly different configurations

Credit to Dave Cunningham for the code, his post in 2017,
https://groups.google.com/g/jsonnet/c/ky6sjYj4UZ0/m/d4lZxWbhAAAJ

Where were you Dave in 2007 when I flunked Programming: FP and Haskell?
I would not have come up with the solution on my own, glad it was shared
in a forum. List comprehensions do not yet feel natural for me to use.